IS-2435: Refactor to using useHistorikk hook
